fix: three Codex findings on the Rust port, and one shared with Go
Verified each against the canonical TypeScript before changing anything.

**A hostile FEN tag overflowed the move counter (Rust only).** PGN spec
9.7 bounds the fullmove field nowhere, so `[FEN "... b - - 0 4294967295"]`
plus one move reached a `u32` addition: a panic in debug, a wrap in
release, where TypeScript and Go both advance to 4294967296. The counter
and `Move::number` are `u64` now, and the increment saturates, so no
input can turn into an `internal` diagnostic. Pinned by a
`numbering.tsv` row, which all three runtimes run.

**Export format accepted a glyph it should refuse (Rust AND Go).** The
two ports range-checked the value where TypeScript narrows the pattern,
so `$0255` read as 255 and `$0000000001` read as 0 and handed `1` back
to the lexer as a move number — a silently wrong parse rather than a
refusal. Both now use the canonical pattern with the trailing `(?!\d)`
spelled as a digit check after the match. The Go half is a pre-existing
divergence this review surfaced rather than anything the port
introduced; fixing it is what lets the four new `strict.tsv` rows pin
the behaviour in all three runtimes at once.

**`san_pattern` was public, and is not a validator.** On its own it
matches the `e2` prefix of `e2e4`; the section 7 boundary that rejects
the pair lives in the matcher. Neither of the other two ports exports
it, so neither does this one. `parse_san` is the public way to ask
whether a string is a move.

Also `.NOTPARALLEL:`, because `build-ts` rewrites the embedded grammar
inside `go/chess.go` and `rs/src/lib.rs` while `build-go` and `build-rs`
read it: under `make -j` the aggregate targets' ordering was advisory.

+++ rs/src/model.rs
@@ -244,8 +244,13 @@ pub struct Move {
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub annotation: Option<Annotation>,
     /// Fullmove number: stated by a move number indication, else counted.
+    ///
+    /// Wide enough that a hostile `FEN` tag cannot overflow the count: PGN
+    /// spec 9.7 puts no bound on that field, and JavaScript counts in
+    /// doubles, so a 32-bit counter would panic where the canonical
+    /// implementation simply keeps going.
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub number: Option<u32>,
+    pub number: Option<u64>,
     /// Side to move: implied by the count, or by a `...` indication.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub side: Option<Side>,
